pyproject.toml
[tool.poetry]
name = "cheeto"
version = "1.2.3"
description = "Utilities for the UC Davis HPC Core Facility."
license = "Proprietary"
authors = [
	"Camille Scott <cswel@ucdavis.edu>"
]
readme = "README.md"
repository = "https://github.com/ucdavis/cheeto"
include = [
	{ path = "cheeto/templates" }
]
packages = [
    { include = "cheeto" }
]

[tool.poetry.scripts]
cheeto = "cheeto.cmds.__main__:main"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
pdbpp = "^0.10.3"

[tool.poetry.requires-plugins]
poetry-bumpversion = "^0.3.3"

[tool.poetry_bumpversion.file."cheeto/__init__.py"]

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.9.2"
ruamel-yaml = "^0.17.21"
marshmallow-dataclass = {extras = ["union"], version = "^8.7.1"}
mergedeep = "^1.3.4"
jinja2 = "^3.1.4"
sh = "^2.0.7"
pytest = "^7.4.4"
filelock = "^3.16.1"
python-ldap = "^3.4.4"
ldap3 = {git = "https://github.com/ucdavis/ldap3.git", rev = "dev"}
gssapi = "^1.9.0"
pymongo = "^4.10.1"
mongoengine = {git = "https://github.com/MongoEngine/mongoengine", rev = "4d3ab60"}
attrs = "^24.2.0"
httpx = "^0.27.0"
python-dateutil = "^2.9.0.post0"
blinker = "^1.8.2"
pyescrypt = "^0.1.0"
xkcdpass = "^1.19.9"
ponderosa = "^0.5.1"
marko = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
